fix: 用户额度字段 int→int64，修复 32 位平台(armv7)大额额度溢出为负数
- User.Quota/UsedQuota/AffQuota/AffHistoryQuota 改 int64(gorm bigint)
- RelayInfo.UserQuota 改 int64，新增 common.GetContextKeyInt64
- logger.FormatQuota 支持 int64
- 预扣费/结算/审计日志等引用点适配
- 测试断言改用 EqualValues 兼容 int/int64

+++ logger/logger.go
@@ -146,7 +146,7 @@ func LogQuota(quota int) string {
 	}
 }
 
-func FormatQuota(quota int) string {
+func FormatQuota(quota int64) string {
 	q := float64(quota)
 	switch operation_setting.GetQuotaDisplayType() {
 	case operation_setting.QuotaDisplayTypeCNY:
